Update getting started document to be more clear
For someone still new to GraphQL, some of the basic concepts or best practices will not be known to them.  While trying to follow along with the documentation, I became a bit confused and had to go looking elsewhere for answers.  I added a few lines to the documentation that would have really helped me as I was working through the steps.